# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.AuthToken:
    properties:
      expiresAt:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Member:
    properties:
      autoStartTerminal:
        type: boolean
      avatar:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      roleKey:
        type: string
      roleType:
        $ref: '#/definitions/models.MemberRole'
      status:
        type: string
      terminalCommand:
        type: string
      terminalPath:
        type: string
      terminalType:
        type: string
      workspaceId:
        type: string
    type: object
  models.MemberCreate:
    properties:
      name:
        type: string
      roleType:
        $ref: '#/definitions/models.MemberRole'
      terminalCommand:
        type: string
      terminalType:
        type: string
    required:
    - name
    - roleType
    type: object
  models.MemberRole:
    enum:
    - owner
    - admin
    - secretary
    - assistant
    - member
    type: string
    x-enum-comments:
      RoleSecretary: 秘书（监工/协调式监督角色）
    x-enum-descriptions:
    - ""
    - ""
    - 秘书（监工/协调式监督角色）
    - ""
    - ""
    x-enum-varnames:
    - RoleOwner
    - RoleAdmin
    - RoleSecretary
    - RoleAssistant
    - RoleMember
  models.Presence:
    properties:
      activity:
        description: '"typing", "viewing", "idle"'
        type: string
      memberId:
        type: string
      targetId:
        description: conversationId or terminalId
        type: string
      targetType:
        description: '"conversation" or "terminal"'
        type: string
      timestamp:
        type: integer
      workspaceId:
        type: string
    type: object
  models.PresenceUpdate:
    properties:
      activity:
        type: string
      targetId:
        type: string
      targetType:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      username:
        type: string
    type: object
  models.UserCreate:
    properties:
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  models.UserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Multi-Agent Collaboration Platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Orchestra API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthToken'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /api/auth/me:
    get:
      description: Get the currently authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthToken'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User registration
      tags:
      - auth
  /api/auth/validate:
    post:
      consumes:
      - application/json
      description: Check if the provided JWT token is valid
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Validate JWT token
      tags:
      - auth
...
  /api/workspaces/{id}/members:
    get:
      description: Get all members in a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Member'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List workspace members
      tags:
      - members
    post:
      consumes:
      - application/json
      description: Create a new member (AI assistant or team member) in a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MemberCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create workspace member
      tags:
      - members
  /api/workspaces/{id}/members/{memberId}:
    delete:
      description: Delete a member from a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: memberId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete workspace member
      tags:
      - members
    get:
      description: Get a single member by ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get workspace member
      tags:
      - members
    put:
      consumes:
      - application/json
      description: Update a member's properties (partial update supported)
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: memberId
        required: true
        type: string
      - description: Member patch data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update workspace member
      tags:
      - members
...
  /api/workspaces/{id}/members/{memberId}/presence:
    post:
      consumes:
      - application/json
      description: Update a member's real-time activity status (typing, viewing, etc.)
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: memberId
        required: true
        type: string
      - description: Presence data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PresenceUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Presence'
      security:
      - BearerAuth: []
      summary: Update member presence
      tags:
      - members
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"